moved kbe.BE.optimize into the class definition
