+++ mypy.ini
@@ -12,6 +12,10 @@
     disallow_untyped_defs = False
     check_untyped_defs = False
 
+[mypy-quemb.shared.external.*]
+    disallow_untyped_defs = False
+    check_untyped_defs = False
+
 
 # TODO: whenever the following packages have stubs available,
 #   stop ignoring them.
